added SearchBar component to make ItemsList more readable
